change field type.

See merge request superid/searcher!91

+++ searchlogic/src/main/java/cn/superid/search/impl/entities/user/user/UserPO.java
@@ -30,7 +30,7 @@ public class UserPO {
   private String[] tags;
   @Field(type = FieldType.Byte)
   private Integer publicType;
-  @Field(type = FieldType.Byte)
+  @Field(type = FieldType.Short)
   private Short infoPublic;
 
   public UserPO() {
